refactor: 🧹 Jules05: Technical debt cleanup — architectural harmonization
- Consolidate institutional 8-layer ATR risk logic into RiskManager.
- Standardize RiskManager and AuditedRiskManager with validate_signal() API.
- Relocate FeatureEngineer to src/data/feature_engineering.py for domain alignment.
- Standardize exports in src/trading/__init__.py and src/core/__init__.py.
- Delete redundant src/trading/risk_engine.py.
- Update main.py trading loop to use unified risk validation and trade tracking.
- Update documentation in docs/technical-debt/DEBT_LOG.md.
