+++ main.py
@@ -27,7 +27,7 @@ if TYPE_CHECKING:
 
     from src.core.audit_log import AuditLogger
     from src.core.decision_support import DecisionSupportSystem
-    from src.core.feature_engineering import FeatureEngineer
+    from src.data.feature_engineering import FeatureEngineer
     from src.core.monitor import Monitor
     from src.core.schemas import TradeSignal
     from src.core.trade_logger import TradeLogger
@@ -414,11 +414,20 @@ def run_live(
                 # 6. Risk approval gate
                 with profile("risk_check"):
                     health = getattr(model, "get_health_metrics", lambda: None)()
-                    risk_approved = (
-                        risk.approve(signal, signal_id=signal_id, model_health=health)
-                        if direction != 0
-                        else False
-                    )
+                    if direction != 0:
+                        # Fetch current positions for institutional exposure checks
+                        open_positions = connector.get_positions(cfg.symbol)
+                        risk_decision = risk.validate_signal(
+                            signal,
+                            market_data=df_raw,
+                            open_positions=open_positions,
+                            model_health=health,
+                        )
+                        risk_approved = risk_decision.is_approved
+                        if risk_approved:
+                            signal.lot_size = risk_decision.adjusted_lot_size
+                    else:
+                        risk_approved = False
 
                 # 7. Execution Filter Cascade
                 filter_decision = None
@@ -573,6 +582,8 @@ def run_live(
                         execution_latency_ms = (time.perf_counter() - execution_start) * 1000
 
                         if ticket:
+                            # Tracking trade count and PnL is now integrated into risk manager hooks
+                            risk.record_trade_open()
                             risk.open_positions[cfg.symbol] = ticket
                             log.info("Order placed", ticket=ticket, latency_ms=execution_latency_ms)
                             if monitor:
