+++ src/trading/risk_manager.py
@@ -3,13 +3,11 @@ MT5 AI/ML Trading Bot - Enterprise Edition
 src/trading/risk_manager.py
 
 Enterprise risk management engine implementing:
+  - ATR-based position sizing (14-period vs 30-day average)
   - Kelly Criterion position sizing (fractional)
+  - Cascading daily loss circuit breakers (Level 1-4)
+  - 8-layer safety cascade signal validation
   - Ray Dalio All-Weather portfolio allocation
-  - Dynamic drawdown protection & circuit breakers
-  - 6-layer entry filter cascade
-
-This module relies on the unified TradeSignal schema from src.core.schemas
-to ensure all signals entering the risk engine are technically valid.
 
 Author : triqbit
 License: MIT
@@ -20,7 +18,9 @@ from __future__ import annotations
 import logging
 from dataclasses import dataclass, field
 from datetime import date
-from typing import Dict, Optional
+from typing import Any, Dict, List, Optional
+
+import pandas as pd
 
 from src.core.config import TradingConfig
 from src.core.monitor import Monitor
@@ -43,6 +43,15 @@ ALLOCATION_WEIGHTS: Dict[str, float] = {
 
 
 @dataclass
+class RiskDecision:
+    """Decision details from the RiskManager."""
+
+    is_approved: bool
+    reason: str = ""
+    adjusted_lot_size: float = 0.0
+
+
+@dataclass
 class DailyStats:
     """Intraday PnL tracker reset each trading day."""
 
@@ -56,7 +65,7 @@ class DailyStats:
 class RiskManager:
     """
     Central risk authority.
-    Every signal must be approved here before reaching the order router.
+    Enforces RISK_LIMITS.md safeguards via an 8-layer cascade.
     """
 
     def __init__(
@@ -70,12 +79,80 @@ class RiskManager:
         self.balance = account_balance
         self.peak_equity = account_balance
         self.daily = DailyStats(peak_equity=account_balance)
-        self.open_positions: Dict[str, int] = {}  # symbol -> ticket
+        self.open_positions: Dict[str, int] = {}  # symbol -> ticket (legacy tracking)
         self.trade_logger = logger_db
         self.monitor = monitor
         logger.info("RiskManager initialised | balance=%.2f", account_balance)
 
     # -- Public API ---------------------------------------------------------
+
+    def validate_signal(
+        self,
+        signal: TradeSignal,
+        market_data: pd.DataFrame,
+        open_positions: List[Dict[str, Any]],
+        model_health: Optional[Dict[str, float]] = None,
+    ) -> RiskDecision:
+        """
+        Validate a trade signal against the 8-layer cascade from RISK_LIMITS.md.
+
+        Layers:
+          1. Circuit Breakers (Equity Drawdown).
+          2. Daily Loss Limits (Level 4 Emergency Stop).
+          3. Activity Limits (Max Daily Trades, Max Consecutive Losses).
+          4. Exposure Limits (Max Concurrent Positions, Single Direction, Total Notional).
+          5. Symbol Allocation (Approved Portfolio).
+          6. Prediction Limits (Min Confidence).
+          7. Risk-Reward Validation (Institutional R:R ratio).
+          8. Model Health (Drift, Accuracy, Calibration).
+        """
+        # Layer 1: Circuit Breakers (Equity Drawdown)
+        if not self._check_circuit_breaker():
+            return RiskDecision(False, "Hard drawdown limit reached")
+
+        # Layer 2: Daily Loss Limits (Level 4)
+        if self.get_daily_loss_level() >= 4:
+            return RiskDecision(False, "Daily loss limit reached (Level 4)")
+
+        # Layer 3: Activity Limits
+        if self.daily.trade_count >= self.cfg.max_trades_per_day:
+            return RiskDecision(False, "Max daily trades reached")
+        if not self._check_consecutive_losses():
+            return RiskDecision(False, "Max consecutive losses reached")
+
+        # Layer 4: Exposure Limits
+        if len(open_positions) >= self.cfg.max_positions:
+            return RiskDecision(False, "Max concurrent positions reached")
+        if not self._check_directional_exposure(signal, open_positions):
+            return RiskDecision(False, "Max directional exposure reached")
+        if not self._check_total_notional(signal, open_positions, market_data):
+            return RiskDecision(False, "Total notional exposure exceeds equity")
+
+        # Layer 5: Symbol Allocation
+        if not self._check_symbol_allocation(signal.symbol):
+            return RiskDecision(False, f"Symbol {signal.symbol} not in approved portfolio")
+
+        # Layer 6: Prediction Limits
+        if not self._check_minimum_confidence(signal.confidence):
+            return RiskDecision(False, f"Confidence {signal.confidence:.2f} too low")
+
+        # Layer 7: Risk-Reward Validation
+        if not self._check_risk_reward(signal):
+            return RiskDecision(False, "Risk-Reward ratio too low")
+
+        # Layer 8: Model Health
+        if not self._check_model_health(model_health):
+            return RiskDecision(False, "Model health metrics below threshold")
+
+        # Calculate final lot size
+        # Default to ATR-based sizing, fallback to Kelly if configured (optional future expansion)
+        adjusted_lots = self.calculate_position_size(signal.symbol, market_data)
+
+        if adjusted_lots < self.cfg.min_lot_size:
+            return RiskDecision(False, f"Calculated lot size {adjusted_lots} below minimum")
+
+        return RiskDecision(True, "Approved", adjusted_lots)
+
     def approve(
         self,
         signal: TradeSignal,
@@ -83,15 +160,15 @@ class RiskManager:
         model_health: Optional[dict] = None,
     ) -> bool:
         """
-        Run the full 8-layer risk filter cascade.
-        Returns True only if ALL layers pass.
+        Legacy 6-layer entry filter cascade.
+        Deprecated: Use validate_signal instead.
         """
         rejection_reason = ""
         if not self._check_circuit_breaker():
             rejection_reason = "Circuit breaker active"
-        elif not self._check_daily_loss():
+        elif not self._check_daily_loss_legacy():
             rejection_reason = "Daily loss limit reached"
-        elif not self._check_max_positions():
+        elif not self._check_max_positions_legacy():
             rejection_reason = "Max positions reached"
         elif not self._check_symbol_allocation(signal.symbol):
             rejection_reason = f"Symbol {signal.symbol} not in portfolio"
@@ -120,6 +197,48 @@ class RiskManager:
                     signal_id=signal_id,
                 )
         return passed
+
+    def calculate_position_size(self, symbol: str, market_data: pd.DataFrame) -> float:
+        """
+        ATR-based position sizing according to RISK_LIMITS.md.
+        """
+        if market_data.empty or "atr" not in market_data.columns:
+            return self.cfg.min_lot_size
+
+        current_atr = market_data["atr"].iloc[-1]
+        # Approx 30 days of M5 if 288 bars/day
+        avg_atr = market_data["atr"].tail(8640).mean()
+
+        vol_multiplier = 1.0
+        ratio = current_atr / avg_atr if avg_atr > 0 else 1.0
+
+        if ratio > self.cfg.volatility_extreme_threshold:
+            return 0.0
+        elif ratio > self.cfg.volatility_very_high_threshold:
+            vol_multiplier = 0.5
+        elif ratio > self.cfg.volatility_high_threshold:
+            vol_multiplier = 0.75
+
+        loss_multiplier = self.get_size_multiplier_from_loss()
+        total_multiplier = vol_multiplier * loss_multiplier
+
+        if total_multiplier <= 0:
+            return 0.0
+
+        # Sizing: risk X% of balance
+        risk_amount = self.balance * self.cfg.risk_per_trade
+        # ATR * 100 converts gold ATR to $ per lot (assumption for XAUUSD)
+        lot_size = (risk_amount / (current_atr * 100)) * total_multiplier
+
+        # Cap at Max Position Size
+        max_notional = self.balance * self.cfg.max_position_size_pct
+        price = market_data["close"].iloc[-1]
+        max_lots = max_notional / (price * 100)
+
+        final_lots = min(lot_size, max_lots)
+        final_lots = max(self.cfg.min_lot_size, round(final_lots, 2))
+
+        return final_lots
 
     def size_position(
         self,
@@ -165,6 +284,10 @@ class RiskManager:
         else:
             self.daily.consecutive_losses = 0
 
+    def record_trade_open(self) -> None:
+        """Increment trade count on open."""
+        self.daily.trade_count += 1
+
     def reset_daily(self) -> None:
         """Must be called at the start of each trading day."""
         if self.monitor:
@@ -172,7 +295,46 @@ class RiskManager:
         self.daily = DailyStats(peak_equity=self.balance)
         logger.info("Daily stats reset")
 
-    # -- Private filter layers ----------------------------------------------
+    # -- Internal layers ----------------------------------------------------
+
+    def _check_circuit_breaker(self) -> bool:
+        if self.peak_equity <= 0:
+            return True
+        drawdown = (self.peak_equity - self.balance) / self.peak_equity
+        if drawdown >= self.cfg.max_drawdown:
+            logger.critical(
+                "CIRCUIT BREAKER: drawdown=%.1f%% - trading halted",
+                drawdown * 100,
+            )
+            if self.trade_logger:
+                self.trade_logger.log_risk_event(
+                    event_type="CIRCUIT_BREAKER",
+                    description=f"Drawdown {drawdown * 100:.1f}% hit {self.cfg.max_drawdown*100}% limit",
+                )
+            if self.monitor:
+                self.monitor.alert_circuit_breaker(drawdown)
+            return False
+        return True
+
+    def get_daily_loss_level(self) -> int:
+        if self.daily.peak_equity <= 0 or self.daily.realised_pnl >= 0:
+            return 0
+        loss_pct = abs(self.daily.realised_pnl) / self.daily.peak_equity
+        if loss_pct >= self.cfg.max_daily_loss:
+            return 4
+        if loss_pct >= self.cfg.daily_loss_lvl3:
+            return 3
+        if loss_pct >= self.cfg.daily_loss_lvl2:
+            return 2
+        if loss_pct >= self.cfg.daily_loss_lvl1:
+            return 1
+        return 0
+
+    def get_size_multiplier_from_loss(self) -> float:
+        level = self.get_daily_loss_level()
+        mapping = {0: 1.0, 1: 1.0, 2: 0.5, 3: 0.25, 4: 0.0}
+        return mapping.get(level, 0.0)
+
     def _check_consecutive_losses(self) -> bool:
         if self.daily.consecutive_losses >= self.cfg.max_losing_streak:
             logger.warning(
@@ -183,7 +345,33 @@ class RiskManager:
             return False
         return True
 
-    def _check_model_health(self, health: Optional[dict]) -> bool:
+    def _check_directional_exposure(
+        self, signal: TradeSignal, open_positions: List[Dict[str, Any]]
+    ) -> bool:
+        net_lots = 0.0
+        for pos in open_positions:
+            vol = pos.get("volume", 0.0)
+            if pos.get("type") == 0:  # BUY
+                net_lots += vol
+            else:  # SELL
+                net_lots -= vol
+
+        net_lots += self.cfg.min_lot_size if signal.direction > 0 else -self.cfg.min_lot_size
+        price_estimate = 2300.0  # Gold estimate
+        notional = abs(net_lots) * price_estimate * 100
+        exposure_pct = notional / self.balance if self.balance > 0 else 1.0
+
+        return exposure_pct <= self.cfg.max_single_direction_pct
+
+    def _check_total_notional(
+        self, signal: TradeSignal, open_positions: List[Dict[str, Any]], market_data: pd.DataFrame
+    ) -> bool:
+        total_lots = sum(pos.get("volume", 0.0) for pos in open_positions) + self.cfg.min_lot_size
+        price = market_data["close"].iloc[-1] if not market_data.empty else 2300.0
+        total_notional = total_lots * price * 100
+        return total_notional < (self.balance * self.cfg.max_total_notional_pct)
+
+    def _check_model_health(self, health: Optional[Dict[str, float]]) -> bool:
         if health is None:
             return True
 
@@ -211,24 +399,7 @@ class RiskManager:
 
         return True
 
-    def _check_circuit_breaker(self) -> bool:
-        drawdown = (self.peak_equity - self.balance) / self.peak_equity
-        if drawdown >= 0.15:  # 15% peak-to-valley kills all trading
-            logger.critical(
-                "CIRCUIT BREAKER: drawdown=%.1f%% - trading halted",
-                drawdown * 100,
-            )
-            if self.trade_logger:
-                self.trade_logger.log_risk_event(
-                    event_type="CIRCUIT_BREAKER",
-                    description=f"Drawdown {drawdown * 100:.1f}% hit 15% limit",
-                )
-            if self.monitor:
-                self.monitor.alert_circuit_breaker(drawdown)
-            return False
-        return True
-
-    def _check_daily_loss(self) -> bool:
+    def _check_daily_loss_legacy(self) -> bool:
         if self.daily.peak_equity == 0:
             return True
         loss_pct = abs(self.daily.realised_pnl) / self.daily.peak_equity
@@ -237,7 +408,7 @@ class RiskManager:
             return False
         return True
 
-    def _check_max_positions(self) -> bool:
+    def _check_max_positions_legacy(self) -> bool:
         if len(self.open_positions) >= self.cfg.max_positions:
             logger.debug("Max positions reached (%d)", self.cfg.max_positions)
             return False
@@ -245,12 +416,13 @@ class RiskManager:
 
     def _check_symbol_allocation(self, symbol: str) -> bool:
         """Block trading on symbols not in the All-Weather portfolio."""
-        if symbol not in ALLOCATION_WEIGHTS:
+        if symbol not in ALLOCATION_WEIGHTS and symbol != self.cfg.symbol:
             logger.warning("Symbol %s not in approved portfolio", symbol)
             return False
         return True
 
-    def _check_minimum_confidence(self, confidence: float, threshold: float = 0.55) -> bool:
+    def _check_minimum_confidence(self, confidence: float) -> bool:
+        threshold = self.cfg.min_confidence
         if confidence < threshold:
             logger.debug("Confidence %.2f below threshold %.2f", confidence, threshold)
             return False
@@ -268,4 +440,4 @@ class RiskManager:
         return True
 
 
-__all__ = ["ALLOCATION_WEIGHTS", "DailyStats", "RiskManager"]
+__all__ = ["ALLOCATION_WEIGHTS", "DailyStats", "RiskDecision", "RiskManager"]
