+++ tests/test_risk_manager_new.py
@@ -2,31 +2,32 @@ import pandas as pd
 import pytest
 
 from src.core.config import TradingConfig
-from src.trading.risk_engine import RiskEngine
+from src.trading.risk_manager import RiskManager
 
 
 @pytest.fixture
-def risk_engine():
+def risk_manager():
     cfg = TradingConfig(MT5_PASSWORD="test", MT5_SERVER="test")
-    return RiskEngine(cfg, 10000.0)
+    return RiskManager(cfg, 10000.0)
 
-def test_drawdown_breaker(risk_engine):
-    risk_engine.update_metrics(6000.0) # 40% drawdown
-    assert not risk_engine._check_drawdown_breaker()
+def test_drawdown_breaker(risk_manager):
+    risk_manager.update_equity(6000.0) # 40% drawdown
+    assert not risk_manager._check_circuit_breaker()
 
-def test_daily_loss_breaker(risk_engine):
-    risk_engine.update_metrics(10000.0, realized_pnl=-600.0) # 6% loss
-    assert risk_engine.get_daily_loss_level() >= 4
+def test_daily_loss_breaker(risk_manager):
+    risk_manager.update_equity(10000.0)
+    risk_manager.record_pnl(-600.0) # 6% loss
+    assert risk_manager.get_daily_loss_level() >= 4
 
-def test_calculate_position_size(risk_engine):
+def test_calculate_position_size(risk_manager):
     data = pd.DataFrame({
         "atr": [1.0] * 100,
         "close": [2300.0] * 100
     })
-    size = risk_engine.calculate_position_size("XAUUSD", data)
+    size = risk_manager.calculate_position_size("XAUUSD", data)
     assert size >= 0.01
 
-def test_validate_signal_rejection(risk_engine):
+def test_validate_signal_rejection(risk_manager):
     from src.core.constants import SignalDirection
     from src.core.schemas import TradeSignal
 
@@ -42,6 +43,6 @@ def test_validate_signal_rejection(risk_engine):
     )
 
     data = pd.DataFrame({"atr": [1.0], "close": [2300.0]})
-    decision = risk_engine.validate_signal(signal, data, [])
+    decision = risk_manager.validate_signal(signal, data, [])
     assert not decision.is_approved
     assert "Confidence" in decision.reason
